Abandoning day 21 part 2!
Made some attempts to speed up but can't see anything that would give an order of magnitude improvement.

Thought I could just work with "number of ON bits" instead of the actual bits. But it doesn't work.

Could run it for ages to get the answer but can't be bothered.